+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- Enable autocomplete for grouping (gorarakelyan)
+
 ## 1.2.2 Nov 27 2020
 - Add circles and interactions on parallel coordinates plot (roubkar)
 - Add parallel coordinates plot chart number and config as title (roubkar)